Launch the API

This function starts the api with the settings it has defined.

Usage

api_run(
  api,
  block = !is_interactive(),
  showcase = is_interactive(),
  ...,
  silent = FALSE
)

api_stop(api)

Arguments

api

A plumber2 api object to launch or stop

block

Should the console be blocked while running (alternative is to run in the background). Defaults to FALSE in interactive sessions and TRUE otherwise.

showcase

Should the default browser open up at the server address. If TRUE then a browser opens at the root of the api, unless the api contains OpenAPI documentation in which case it will open at that location. If a string the string is used as a path to add to the root before opening.

...

Arguments passed on to the start handler

silent

Should startup messaging by silenced

Value

These functions return the api object allowing for easy chaining with the pipe, even though they will often be the last part of the chain